// src/app/enums/race.enum.ts
export enum Race {
  // Alliance races
  Human = 'Human',
  NightElf = 'Night Elf',
  Dwarf = 'Dwarf',
  Gnome = 'Gnome',
  Draenei = 'Draenei',
  Worgen = 'Worgen',
  VoidElf = 'Void Elf',
  LightforgedDraenei = 'Lightforged Draenei',
  DarkIronDwarf = 'Dark Iron Dwarf',
  KulTiran = 'Kul Tiran',
  Mechagnome = 'Mechagnome',

  // Horde races
  Orc = 'Orc',
  Undead = 'Undead (Forsaken)',
  Tauren = 'Tauren',
  Troll = 'Troll',
  BloodElf = 'Blood Elf',
  Goblin = 'Goblin',
  Nightborne = 'Nightborne',
  HighmountainTauren = 'Highmountain Tauren',
  MagharOrc = "Mag'har Orc",
  ZandalariTroll = 'Zandalari Troll',
  Vulpera = 'Vulpera',

  // Neutral races (can choose faction)
  Pandaren = 'Pandaren',
  Dracthyr = 'Dracthyr'
}